infra: new test
