+++ src/dist-indexer.js
