+++ src/ls-types.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-const fs = require('fs')
+const fs = require('node:fs')
 const transformFilename = require('./transform-filename')
 
 function lsTypes (dir, dones, callback) {
